+++ CHANGELOG.md
@@ -22,6 +22,11 @@ Changelog
   * DELETE `/1.1/users/<user_id>/schedules/<schedule_id>`
   * PUT `/1.1/users/<user_id>/schedules/<schedule_id>`
 
+* A new API for associating an group with a schedule has been added:
+
+  * DELETE `/1.1/groups/<group_id>/schedules/<schedule_id>`
+  * PUT `/1.1/groups/<group_id>/schedules/<schedule_id>`
+
 
 17.13
 -----